requirements.txt
openai
numpy
pandas
google
streamlit
python-dotenv
newspaper3k
langchain
tiktoken
lxml[html_clean]
openpyxl